tests/wpt/css-tests/css-transitions-1_dev/xhtml1/transition-property-002.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
        <meta charset="utf-8" />
        <title>CSS Transitions Test: Parsing invalid transition-property</title>
        <meta content="Test checks that unrecognized or non-animatable properties must be kept in the list to preserve the matching of indices." name="assert" />
        <link href="http://www.w3.org/TR/css3-transitions/#transition-property-property" rel="help" title="2.1. The 'transition-property' Property" />
        <link href="http://rodneyrehm.de/en/" rel="author" title="Rodney Rehm" />
        <meta content="dom " name="flags" />

        <script src="/resources/testharness.js" type="text/javascript"></script>
        <script src="/resources/testharnessreport.js" type="text/javascript"></script>

        <script src="./support/vendorPrefix.js" type="text/javascript"></script>
        <script src="./support/helper.js" type="text/javascript"></script>

        <script id="metadata_cache">/*
        {
          "parse 'none, all'": {},
          "parse 'all, none'": {},
          "parse 'foobar'": {},
          "parse 'all, foobar'": {},
          "parse 'foobar, all'": {},
          "parse 'inherit'": {},
          "parse 'initial'": {}
        }
        */</script>
    </head>
    <body>
        <!-- required by testharnessreport.js -->
        <div id="log"></div>
        <!-- elements used for testing -->
        <div id="container">
            <div id="transition"></div>
        </div>

        <script>
            var transition = document.getElementById('transition');

            // syntax: none | [ all | &lt;IDENT&gt; ] [ ‘,’ [ all | &lt;IDENT&gt; ] ]*
            var values = {
                'none, all' : 'none, all',
                'all, none' : 'all, none',
                'foobar' : 'foobar',
                'all, foobar' : 'all, foobar',
                'foobar, all' : 'foobar, all',
                'inherit' : 'padding',
                'initial' : 'all'
            };

            for (var key in values) {
                test(function() {
                    setStyle({
                        '#container': {'transition-property': 'padding'},
                        '#transition': {'transition-property': key}
                    });
                    var result = computedStyle(transition, 'transition-property');
                    assert_equals(result, values[key]);
                }, "parse '" + key + "'");
            }
        </script>
    

</body></html>
